Add logic to load ssl certificates, https server, and have websocket server run ontop of https server
