+++ app/public/constants.js
@@ -1,5 +1,4 @@
-const WEBSOCKET_PORT = 443;
-const ELASTIC_IP_ADDRESS = "54.172.133.168";
+const DOMAIN_NAME = "cyberspacesaboteur.com"
 const IS_PROD = false;
 
 const ROLES = {
