+++ app/public/index.html
@@ -16,6 +16,7 @@
     </div>
 </body>
 <script src="./addStars.js"></script>
+<script src="./constants.js"></script>
 <script src="./router.js"></script>
 <script src="./homepage.js"></script>
 </html>
