+++ app/public/login/index.html
@@ -96,6 +96,7 @@
       </form>
     </div>
   </body>
+  <script src="../constants.js"></script>
   <script src="../router.js"></script>
   <script src="./loginpage.js"></script>
 </html>
