+++ app/public/websocket.js
@@ -1,8 +1,7 @@
 const WS = (() => {
     
-    const WEBSOCKET_ORIGIN = IS_PROD ? ELASTIC_IP_ADDRESS : "localhost";
-    const webSocket = new WebSocket(`ws://${WEBSOCKET_ORIGIN}:${WEBSOCKET_PORT}/`);
-    console.log("WS - Websocket connection opened");
+    const WEBSOCKET_URL = IS_PROD ? `wss://${DOMAIN_NAME}` : "ws://localhost:80";
+    const webSocket = new WebSocket(WEBSOCKET_URL);
 
     const send = (data) => {
         webSocket.send(JSON.stringify(data));
