+++ env.json
@@ -1,8 +1,8 @@
 {
-    "WEBSOCKET_PORT" : 443,
+    "WEBSOCKET_PORT" : 80,
     "PROD_HOSTNAME" : "0.0.0.0",
     "LOCAL_HOSTNAME" : "localhost",
-    "PROD_PORT" : 80,
+    "PROD_PORT" : 443,
     "LOCAL_PORT" : 3000,
     "IS_PROD" : false
 }
