+++ package.json
@@ -17,6 +17,8 @@
     "cookie-parser": "^1.4.6",
     "crypto": "^1.0.1",
     "express": "^4.18.2",
+    "fs": "^0.0.1-security",
+    "https": "^1.0.0",
     "pg": "^8.11.1",
     "pm2": "^5.3.0",
     "ws": "^8.13.0"
